Feat : 단어퀴즈 답안 제출 기능

+++ backend/src/main/java/com/ieng/ieng/domain/word/repository/WordRepository.java
@@ -7,5 +7,5 @@ import java.util.List;
 
 public interface WordRepository extends JpaRepository<Word,Long> {
     List<Word> findAll();
-    Word findWordByWordSequence(int wordSequence);
+    Word findWordByWordSequence(Long wordSequence);
 }
